Fix dependency loop
