+++ src/views/Game/GameDock.tsx
@@ -29,7 +29,8 @@ import { openGameLogModal } from "./GameLogModal";
 import { sfx } from "sfx";
 import { alert } from "swal_config";
 import { challengeFromBoardPosition } from "ChallengeModal/ForkModal";
-import { errorAlerter, doAnnul } from "misc";
+import { errorAlerter } from "misc";
+import { doAnnul } from "moderation";
 import { openReport } from "Report";
 import { game_control } from "./game_control";
 import { openGameInfoModal } from "./GameInfoModal";
